utils(store): move static store implementation into `hash_utils::store`
