+++ compiler/hash-exhaustiveness/src/deconstruct.rs
@@ -12,11 +12,14 @@ use std::{
 use hash_intrinsics::utils::PrimitiveUtils;
 use hash_tir::{
     data::{CtorDefId, DataTy},
-    environment::stores::StoreId,
     pats::PatId,
     tys::{Ty, TyId},
 };
-use hash_utils::{itertools::Itertools, smallvec::SmallVec, store::Store};
+use hash_utils::{
+    itertools::Itertools,
+    smallvec::SmallVec,
+    store::{statics::StoreId, Store},
+};
 
 use super::{construct::DeconstructedCtor, fields::Fields};
 use crate::{
