+++ compiler/hash-intrinsics/src/primitives.rs
@@ -3,17 +3,14 @@ use std::iter::once;
 
 use hash_tir::{
     data::{ArrayCtorInfo, DataDefId, NumericCtorBits, NumericCtorInfo, PrimitiveCtorInfo},
-    environment::{
-        env::{AccessToEnv, Env},
-        stores::SequenceStoreValue,
-    },
+    environment::env::{AccessToEnv, Env},
     mods::{ModMemberData, ModMemberValue},
     params::{Param, ParamData},
     symbols::sym,
     tys::Ty,
     utils::{common::CommonUtils, AccessToUtils},
 };
-use hash_utils::store::Store;
+use hash_utils::store::{statics::SequenceStoreValue, Store};
 
 macro_rules! defined_primitives {
     ($($name:ident),* $(,)?) => {
