+++ compiler/hash-tir/src/environment/stores.rs
@@ -1,12 +1,5 @@
 // @@Docs
-use std::{
-    marker::PhantomData,
-    ops::{Deref, DerefMut},
-    sync::OnceLock,
-};
-
-use hash_utils::store::sequence::{SequenceStore, SequenceStoreInternalData, SequenceStoreKey};
-use parking_lot::RwLock;
+use std::sync::OnceLock;
 
 use super::super::{
     args::{ArgsStore, PatArgsStore},
@@ -90,346 +83,12 @@ pub fn tir_stores() -> &'static Stores {
     STORES.get_or_init(Stores::new)
 }
 
-/// A trait for a store ID which can be used to access a store in a store.
-pub trait StoreId: Sized + Copy {
-    type Value;
-    type ValueRef: ?Sized;
-    type ValueBorrow: Deref<Target = Self::ValueRef>;
-    type ValueBorrowMut: DerefMut<Target = Self::ValueRef>;
-
-    /// Borrow the value associated with this ID.
-    fn borrow(self) -> Self::ValueBorrow;
-
-    /// Borrow the value associated with this ID mutably.
-    fn borrow_mut(self) -> Self::ValueBorrowMut;
-
-    /// Get the value associated with this ID.
-    fn value(self) -> Self::Value;
-
-    /// Map the value associated with this ID to a new value.
-    fn map<R>(self, f: impl FnOnce(&Self::ValueRef) -> R) -> R;
-
-    /// Modify the value associated with this ID.
-    fn modify<R>(self, f: impl FnOnce(&mut Self::ValueRef) -> R) -> R;
-
-    /// Set the value associated with this ID.
-    fn set(self, value: Self::Value);
-}
-
-/// A trait for a sequence store ID which can be used to access a store in
-/// `STORES`.
-pub trait SequenceStoreValue: Sized {
-    type Id: StoreId;
-    type ElementId: StoreId;
-
-    /// Create a new empty value in the store.
-    fn empty_seq() -> Self::Id;
-
-    /// Create a new value in the store from the given iterator of functions.
-    fn seq<F: FnOnce(Self::ElementId) -> Self, I: IntoIterator<Item = F>>(iter: I) -> Self::Id
-    where
-        I::IntoIter: ExactSizeIterator;
-}
-
-/// A trait for a store ID containing single items which can be used to access a
-/// store in `STORES`.
-pub trait SingleStoreValue: Sized {
-    type Id: StoreId;
-
-    /// Create a new value in the store from the given function.
-    fn create(self) -> Self::Id {
-        Self::create_with(|_| self)
-    }
-
-    /// Create a new value in the store from the given function.
-    fn create_with<F: FnOnce(Self::Id) -> Self>(value: F) -> Self::Id;
-}
-
-#[derive(Debug)]
-pub struct DefaultIndirectSequenceStore<K, V> {
-    data: SequenceStoreInternalData<V>,
-    _phantom: PhantomData<K>,
-}
-
-impl<K, V> Default for DefaultIndirectSequenceStore<K, V> {
-    fn default() -> Self {
-        Self { data: RwLock::new(Vec::new()), _phantom: PhantomData }
-    }
-}
-
-impl<K, V> DefaultIndirectSequenceStore<K, V> {
-    pub fn new() -> Self {
-        Self::default()
-    }
-}
-
-impl<K: SequenceStoreKey<ElementKey = V>, V: Clone> SequenceStore<K, V>
-    for DefaultIndirectSequenceStore<K, V>
-{
-    fn internal_data(&self) -> &SequenceStoreInternalData<V> {
-        &self.data
-    }
-}
-
-/// Automatically implement `StoreId` and `SequenceStoreId` for a sequence store
-/// ID type.
-#[macro_export]
-macro_rules! static_sequence_store_indirect {
-    (store = $store_vis:vis $store:ident, id = $id_vis:vis $id:ident[$el_id:ident], store_name = $store_name:ident, store_source = $store_source:expr) => {
-        $store_vis type $store = $crate::environment::stores::DefaultIndirectSequenceStore<$id, $el_id>;
-        hash_utils::new_sequence_store_key_indirect!($id_vis $id, $el_id);
-
-        impl $crate::environment::stores::StoreId for $id {
-            type Value = Vec<$el_id>;
-            type ValueRef = [$el_id];
-            type ValueBorrow = hash_utils::store::sequence::SequenceStoreBorrowHandle<'static, [$el_id]>;
-            type ValueBorrowMut = hash_utils::store::sequence::SequenceStoreBorrowMutHandle<'static, [$el_id]>;
-
-            fn borrow(self) -> Self::ValueBorrow {
-                $store_source.$store_name().borrow(self)
-            }
-
-            fn borrow_mut(self) -> Self::ValueBorrowMut {
-                $store_source.$store_name().borrow_mut(self)
-            }
-
-            fn value(self) -> Self::Value {
-                $store_source.$store_name().get_vec(self)
-            }
-
-            fn map<R>(self, f: impl FnOnce(&Self::ValueRef) -> R) -> R {
-                $store_source.$store_name().map_fast(self, f)
-            }
-
-            fn modify<R>(self, f: impl FnOnce(&mut Self::ValueRef) -> R) -> R {
-                $store_source.$store_name().modify_fast(self, f)
-            }
-
-            fn set(self, value: Self::Value) {
-                $store_source.$store_name().set_from_slice_cloned(self, &value);
-            }
-        }
-
-        impl std::fmt::Debug for $id {
-            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-                f.debug_list().entries(self.value().iter()).finish()
-            }
-        }
-
-        use $crate::environment::stores::StoreId;
-
-        impl From<($id, usize)> for $el_id {
-            fn from((id, index): ($id, usize)) -> Self {
-                $id::from(id).map(|value| value[index].clone())
-            }
-        }
-    };
-}
-
-/// Automatically implement `StoreId` and `SequenceStoreId` for a sequence store
-/// ID type.
-#[macro_export]
-macro_rules! static_sequence_store_direct {
-    (
-        store = $store_vis:vis $store:ident,
-        id = $id_vis:vis $id:ident[$el_id:ident],
-        value = $value:ty,
-        store_name = $store_name:ident,
-        store_source = $store_source:expr,
-        derives = Debug
-    ) => {
-        static_sequence_store_direct! {
-            store = $store_vis $store,
-            id = $id_vis $id[$el_id],
-            value = $value,
-            store_name = $store_name,
-            store_source = $store_source
-        }
-        hash_utils::impl_debug_for_sequence_store_element_key!($el_id);
-    };
-    (
-        store = $store_vis:vis $store:ident,
-        id = $id_vis:vis $id:ident[$el_id:ident],
-        value = $value:ty,
-        store_name = $store_name:ident,
-        store_source = $store_source:expr
-        $(, derives = $($extra_derives:ident),*)?
-    ) => {
-        $store_vis type $store = hash_utils::store::sequence::DefaultSequenceStore<$id, $value>;
-        hash_utils::new_sequence_store_key_direct!($id_vis $id, $el_id $(, el_derives = [$($extra_derives),*])?);
-
-        impl $crate::environment::stores::StoreId for $id {
-            type Value = Vec<$value>;
-            type ValueRef = [$value];
-            type ValueBorrow = hash_utils::store::sequence::SequenceStoreBorrowHandle<'static, [$value]>;
-            type ValueBorrowMut = hash_utils::store::sequence::SequenceStoreBorrowMutHandle<'static, [$value]>;
-
-            fn borrow(self) -> Self::ValueBorrow {
-                $store_source.$store_name().borrow(self)
-            }
-
-            fn borrow_mut(self) -> Self::ValueBorrowMut {
-                $store_source.$store_name().borrow_mut(self)
-            }
-
-            fn value(self) -> Self::Value {
-                $store_source.$store_name().get_vec(self)
-            }
-
-            fn map<R>(self, f: impl FnOnce(&Self::ValueRef) -> R) -> R {
-                $store_source.$store_name().map_fast(self, f)
-            }
-
-            fn modify<R>(self, f: impl FnOnce(&mut Self::ValueRef) -> R) -> R {
-                $store_source.$store_name().modify_fast(self, f)
-            }
-
-            fn set(self, value: Self::Value) {
-                $store_source.$store_name().set_from_slice_cloned(self, &value);
-            }
-        }
-
-        impl std::fmt::Debug for $id {
-            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-                use hash_utils::store::sequence::TrivialSequenceStoreKey;
-                let entries: Vec<_> = self.iter().collect();
-                f.debug_tuple(stringify!($id)).field(&self.index).field(&self.len)
-                    .field(&entries)
-                    .finish()
-            }
-        }
-
-        impl $crate::environment::stores::SequenceStoreValue for $value {
-            type Id = $id;
-            type ElementId = $el_id;
-
-            fn empty_seq() -> Self::Id {
-                $store_source.$store_name().create_from_slice(&[])
-            }
-
-            fn seq<F: FnOnce($el_id) -> Self, I: IntoIterator<Item = F>>(values: I) -> Self::Id
-            where
-                I::IntoIter: ExactSizeIterator,
-            {
-                $store_source.$store_name().create_from_iter_with(values)
-            }
-        }
-
-        impl $crate::environment::stores::StoreId for $el_id {
-            type Value = $value;
-            type ValueRef = $value;
-            type ValueBorrow = hash_utils::store::sequence::SequenceStoreBorrowHandle<'static, $value>;
-            type ValueBorrowMut = hash_utils::store::sequence::SequenceStoreBorrowMutHandle<'static, $value>;
-
-            fn borrow(self) -> Self::ValueBorrow {
-                use hash_utils::store::sequence::TrivialKeySequenceStore;
-                $store_source.$store_name().borrow_element(self)
-            }
-
-            fn borrow_mut(self) -> Self::ValueBorrowMut {
-                use hash_utils::store::sequence::TrivialKeySequenceStore;
-                $store_source.$store_name().borrow_element_mut(self)
-            }
-
-            fn value(self) -> Self::Value {
-                use hash_utils::store::sequence::TrivialKeySequenceStore;
-                $store_source.$store_name().get_element(self.into())
-            }
-
-            fn map<R>(self, f: impl FnOnce(&Self::ValueRef) -> R) -> R {
-                $store_source.$store_name().map_fast(self.0, |v| f(&v[self.1]))
-            }
-
-            fn modify<R>(self, f: impl FnOnce(&mut Self::ValueRef) -> R) -> R {
-                $store_source.$store_name().modify_fast(self.0, |v| f(&mut v[self.1]))
-            }
-
-            fn set(self, value: Self::Value) {
-                $store_source.$store_name().set_at_index(self.0, self.1, value);
-            }
-        }
-    };
-}
-
-/// Automatically implement `StoreId` and `SingleStoreId` for a single store ID
-/// type.
-#[macro_export]
-macro_rules! static_single_store {
-    (
-        store = $store_vis:vis $store:ident,
-        id = $id_vis:vis $id:ident,
-        value = $value:ty,
-        store_name = $store_name:ident,
-        store_source = $store_source:expr,
-        derives = Debug
-    ) => {
-        static_single_store! {
-            store = $store_vis $store,
-            id = $id_vis $id,
-            value = $value,
-            store_name = $store_name,
-            store_source = $store_source
-        }
-        hash_utils::impl_debug_for_store_key!($id);
-    };
-    (
-        store = $store_vis:vis $store:ident,
-        id = $id_vis:vis $id:ident,
-        value = $value:ty,
-        store_name = $store_name:ident,
-        store_source = $store_source:expr
-        $(, derives = $($extra_derives:ident),*)?) => {
-        $store_vis type $store = hash_utils::store::DefaultStore<$id, $value>;
-        hash_utils::new_store_key!($id_vis $id $(, derives = $($extra_derives),*)?);
-
-        impl $crate::environment::stores::StoreId for $id {
-            type Value = $value;
-            type ValueRef = $value;
-            type ValueBorrow = hash_utils::store::StoreBorrowHandle<'static, $value>;
-            type ValueBorrowMut = hash_utils::store::StoreBorrowMutHandle<'static, $value>;
-
-            fn borrow(self) -> Self::ValueBorrow {
-                $store_source.$store_name().borrow(self)
-            }
-
-            fn borrow_mut(self) -> Self::ValueBorrowMut {
-                $store_source.$store_name().borrow_mut(self)
-            }
-
-            fn value(self) -> Self::Value {
-                use hash_utils::store::CloneStore;
-                $store_source.$store_name().get(self)
-            }
-
-            fn map<R>(self, f: impl FnOnce(&Self::Value) -> R) -> R {
-                $store_source.$store_name().map_fast(self, f)
-            }
-
-            fn modify<R>(self, f: impl FnOnce(&mut Self::Value) -> R) -> R {
-                $store_source.$store_name().modify_fast(self, f)
-            }
-
-            fn set(self, value: Self::Value) {
-                use hash_utils::store::CloneStore;
-                $store_source.$store_name().set(self, value);
-            }
-        }
-
-        impl $crate::environment::stores::SingleStoreValue for $value {
-            type Id = $id;
-            fn create_with<F: FnOnce(Self::Id) -> Self>(value: F) -> Self::Id {
-                $store_source.$store_name().create_with(value)
-            }
-        }
-    };
-}
-
 #[macro_export]
 macro_rules! tir_debug_value_of_sequence_store_element_id {
     ($id:ident) => {
         impl std::fmt::Debug for $id {
             fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-                use $crate::environment::stores::StoreId;
+                use hash_utils::store::statics::StoreId;
                 f.debug_tuple(stringify!($id))
                     .field(&(&self.0.index, &self.0.len))
                     .field(&self.1)
@@ -445,7 +104,7 @@ macro_rules! tir_debug_value_of_single_store_id {
     ($id:ident) => {
         impl std::fmt::Debug for $id {
             fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-                use $crate::environment::stores::StoreId;
+                use hash_utils::store::statics::StoreId;
                 f.debug_tuple(stringify!($id)).field(&self.index).field(&self.value()).finish()
             }
         }
@@ -457,7 +116,7 @@ macro_rules! tir_debug_name_of_store_id {
     ($id:ident) => {
         impl std::fmt::Debug for $id {
             fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-                use $crate::environment::stores::StoreId;
+                use hash_utils::store::statics::StoreId;
                 f.debug_tuple(stringify!($id)).field(&self.value().name).finish()
             }
         }
@@ -467,6 +126,6 @@ macro_rules! tir_debug_name_of_store_id {
 #[macro_export]
 macro_rules! tir_get {
     ($id:expr, $member:ident) => {{
-        $crate::environment::stores::StoreId::map($id, |x| x.$member)
+        hash_utils::store::statics::StoreId::map($id, |x| x.$member)
     }};
 }
