+++ compiler/hash-typecheck/src/unification.rs
@@ -7,7 +7,6 @@ use hash_tir::{
     args::ArgsId,
     context::ScopeKind,
     data::DataDefCtors,
-    environment::stores::StoreId,
     fns::{FnCallTerm, FnTy},
     holes::Hole,
     lits::Lit,
@@ -18,7 +17,9 @@ use hash_tir::{
     tys::{Ty, TyId},
     utils::{common::CommonUtils, traversing::Atom},
 };
-use hash_utils::store::{CloneStore, SequenceStoreKey, Store, TrivialSequenceStoreKey};
+use hash_utils::store::{
+    statics::StoreId, CloneStore, SequenceStoreKey, Store, TrivialSequenceStoreKey,
+};
 use once_cell::unsync::OnceCell;
 
 use crate::{
