+++ compiler/hash-utils/src/store/mod.rs
@@ -2,10 +2,11 @@
 //! efficient way, with interior mutability.
 
 pub mod base;
+pub use base::*;
+
 pub mod partial;
 pub mod sequence;
 pub mod statics;
 
-pub use base::*;
 pub use partial::*;
 pub use sequence::*;
